refactor(user-auth): split code in 2 files This commit takes all the resolvers that update the user credentials and puts them in another file
